fix some bug for training

+++ safebench/gym_carla/env_wrapper.py
@@ -205,7 +205,7 @@ class ObservationWrapper(gym.Wrapper):
         elif self.agent_obs_type == 'ego_state':
             process_obs = obs['ego_state']
         elif self.agent_obs_type == 'plant':
-            process_obs = obs['plant_encoded_state']
+            process_obs = obs['plant_encoded_state'].astype(np.float64)
         elif self.agent_obs_type == 'no_obs':
             process_obs = obs
         else:
